requirements.txt
streamlit>=1.28.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
chromadb>=0.4.0
pypdf>=3.0.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
rich>=13.0.0
tiktoken>=0.5.0